GalleryGridCard.tsx renamed to GalleryCardGrid.tsx
